// FoodDiary.Web.Client/src/app/shared/api/sdk/generated/model/dashboard-snapshot-http-response.ts
/**
 * FoodDiary User API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { CycleHttpResponse } from './cycle-http-response';
import { DailyAdviceHttpResponse } from './daily-advice-http-response';
import { DailyCaloriesHttpResponse } from './daily-calories-http-response';
import { DashboardLayoutHttpModel } from './dashboard-layout-http-model';
import { DashboardMealsHttpResponse } from './dashboard-meals-http-response';
import { DashboardStatisticsHttpResponse } from './dashboard-statistics-http-response';
import { DashboardWaistHttpResponse } from './dashboard-waist-http-response';
import { DashboardWeightHttpResponse } from './dashboard-weight-http-response';
import { FastingSessionHttpResponse } from './fasting-session-http-response';
import { HydrationDailyHttpResponse } from './hydration-daily-http-response';
import { TdeeInsightHttpResponse } from './tdee-insight-http-response';
import { WaistEntrySummaryHttpResponse } from './waist-entry-summary-http-response';
import { WeightEntrySummaryHttpResponse } from './weight-entry-summary-http-response';

export interface DashboardSnapshotHttpResponse {
    advice?: DailyAdviceHttpResponse | null;
    caloriesBurned?: number;
    currentCycle?: CycleHttpResponse | null;
    currentFastingSession?: FastingSessionHttpResponse | null;
    dailyGoal?: number;
    dashboardLayout?: DashboardLayoutHttpModel | null;
    date?: string;
    dateTo?: string;
    hydration?: HydrationDailyHttpResponse | null;
    meals?: DashboardMealsHttpResponse;
    statistics?: DashboardStatisticsHttpResponse;
    tdeeInsight?: TdeeInsightHttpResponse | null;
    waist?: DashboardWaistHttpResponse;
    waistTrend?: Array<WaistEntrySummaryHttpResponse> | null;
    weeklyCalorieGoal?: number;
    weeklyCalories?: Array<DailyCaloriesHttpResponse>;
    weight?: DashboardWeightHttpResponse;
    weightTrend?: Array<WeightEntrySummaryHttpResponse> | null;
}